+++ examples/C++/main.cpp
@@ -2,6 +2,7 @@
 #include <SHARPlib/constants.h>
 #include <SHARPlib/layer.h>
 #include <SHARPlib/parcel.h>
+#include <SHARPlib/peters_lifter.h>
 #include <SHARPlib/thermo.h>
 #include <SHARPlib/winds.h>
